requirements.txt
streamlit==1.32.0
pyyaml==6.0.1
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
matplotlib==3.8.2 